WIP - Move out printing related code from JsonObject
