+++ modules/core/shared/src/main/scala/io/circe/package.scala
@@ -1,5 +1,4 @@
 package io
 
 package object circe {
-  val Printer = io.circe.prin
 }
