+++ modules/printer/shared/src/main/scala/io/circe/printer/package.scala
@@ -5,10 +5,9 @@ import java.lang.StringBuilder
 import java.nio.charset.{Charset, StandardCharsets}
 import java.nio.{ByteBuffer, CharBuffer}
 import java.util.concurrent.CopyOnWriteArrayList
-
 import io.circe._
 import io.circe.{Printer => CorePrinter}
-
+import io.circe.JsonObject.{LinkedHashMapJsonObject, MapAndVectorJsonObject}
 import scala.annotation.switch
 
 
@@ -355,7 +354,11 @@ package object printer {
         }
       }
 
-      final def onObject(value: JsonObject): Unit = value.appendToFolder(this)
+      final def onObject(value: JsonObject): Unit = {
+        case v: LinkedHashMapJsonObject => JsonObjectPrinter.appendToFolder(v, this)
+        case v: MapAndVectorJsonObject => JsonObjectPrinter.appendToFolder(v, this)
+      }
+
     }
 
     private[circe] final case class Pieces(
